client/requirements.txt
pynput
Pillow
winocr
requests
